refactor(db): extract helpers (uuid, fts, formatting) to db/src/helpers.rs
